Changes to get odom data with correct time stamp for cartographer
In my ROS2 system the odom topic is published by the microcontroller but the time stamp that publishes it with is not correct and same as the one on my desktop. Due to this I changed the code in the microcontroller, pico in this case, to publish the topic odom_raw instead of odom. This odom_raw is then consumed by another node that runs on the host pc ( raspberry pi or desktop) and publishes the topic odom with the correct timestamp that is based on the laser topic( which is published from lidar) and base_link transformations. This is done to make sure that the cartographer works and there is no mismatch in time stamp.